feat: combine db service and models
